Refator: sign up, sign in 서로 변경
sign up과 sign in을 모든 파일에 반대로 적어놓아서 두 이름 모두 수정하였습니다.
