+++ index.html
@@ -14,8 +14,8 @@
   <body>
     <header>
       <div>
-        <a href="/shared.html"><img src="./assets/Linkbrary.png"></a>
-        <a href="/signup.html"><button id="LoginBtn"type="button">로그인</button> </a>
+        <a href="/"><img src="./assets/Linkbrary.png"></a>
+        <a href="/signin.html"><button id="LoginBtn"type="button">로그인</button> </a>
       </div>
     </header>
     <section class="section-top">
@@ -27,7 +27,7 @@
             쉽게 저장하고 <br class="line-break">관리해 보세요.
           </p>
         </div>
-        <a href="/signin.html"><button id="addLinkBtn" type="button">링크 추가하기</button></a>
+        <a href="/signup.html"><button id="addLinkBtn" type="button">링크 추가하기</button></a>
       </div>
       <!-- headline 아래 cards 이미지 -->
       <div class="cards-img">
